4 Integration Tests for Search feature
First test searches for a user and goes to their profile

Second test tries to search "Poop" and it ensures that the test account does not show up

Third test uses the "Endorsed" filter, searches for a user and then goes to their profile

Fourth test uses the "Endorsed" filter and searches for the test account. Ensures that it should not appear because the test account is not endorsed.